# swagger-docs/swagger.json
{
    "swagger": "2.0",
    "paths": {
...
        "/orderschedule/index.php/Home/Roles/create": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "\u521b\u5efa\u89d2\u8272",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "roleName",
                        "in": "\u5fc5\u4f20",
                        "description": "\u89d2\u8272\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Roles/detail": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "\u83b7\u53d6\u5355\u4e2a\u89d2\u8272",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "\u5fc5\u4f20",
                        "description": "\u89d2\u8272id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Roles/delete": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "\u5220\u9664\u89d2\u8272",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "\u5fc5\u4f20",
                        "description": "\u89d2\u8272id",
                        "required": true,
                        "type": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Roles/update": {
            "patch": {
                "tags": [
                    "Roles"
                ],
                "summary": "\u66f4\u65b0\u89d2\u8272",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "\u5fc5\u4f20",
                        "description": "\u89d2\u8272id",
                        "required": true,
                        "type": "int"
                    },
                    {
                        "name": "roleName",
                        "in": "\u5fc5\u4f20",
                        "description": "\u89d2\u8272\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
    },
...
}